Create Conditionals
The user will be prompted to put in a number and the program will determine if it is prime or not. Then the user will be prompted to input a password; any password that is created that is less than 6 letters the output will prompt will say add more letters and if they do not have a number in the password the output will be Password must contain at least one numeric value within the password. If all of these conditions are met the user will get an output of Your set!